Revert moving MutableColumn to @grafana/data
